Refactor fixture import statements in test code
The import statements for database cursor and dev_db_connection fixtures in multiple test files have been updated after the fixtures.py file was moved from 'tests/middleware' directory to 'tests' directory. The refactor ensures test modules correctly reference the fixtures from their new location.
