+++ tests/middleware/test_archives_queries.py
@@ -6,10 +6,9 @@ from middleware.archives_queries import (
     ARCHIVES_GET_COLUMNS,
 )
 from tests.middleware.helper_functions import (
-    insert_test_agencies_and_sources,
     has_expected_keys,
 )
-from tests.middleware.fixtures import (
+from tests.fixtures import (
     dev_db_connection,
     db_cursor,
     connection_with_test_data,
