+++ tests/middleware/test_login_queries.py
@@ -9,7 +9,7 @@ from middleware.login_queries import (
     is_admin,
 )
 from tests.middleware.helper_functions import create_test_user
-from tests.middleware.fixtures import dev_db_connection, db_cursor
+from tests.fixtures import db_cursor, dev_db_connection
 
 
 def test_login_query(db_cursor: psycopg2.extensions.cursor) -> None:
